第一个copote OA release版本 userid改为了long,数据库bigint

+++ src/person/daizhongde/authority/hibernate/dao/TAuthorityModuleDAO.java
@@ -65,7 +65,7 @@ public interface TAuthorityModuleDAO extends SpringHibernateDao{
 	 * @param userid
 	 * @return
 	 */
-	public abstract List findByUserId(Integer userId);
+	public abstract List findByUserId(Long userId);
 	
 	public abstract List findChildrenRecursive( Integer parentId );
 	/**
