+++ src/person/daizhongde/authority/hibernate/dao/TAuthorityRbrelationDAO.java
@@ -17,7 +17,7 @@ public interface TAuthorityRbrelationDAO extends SpringHibernateDao {
 	 * @param userId
 	 * @return
 	 */
-	public abstract List findUserButton4tb(Integer moduleId, Integer userId  );
+	public abstract List findUserButton4tb(Integer moduleId, Long userId  );
 		
 	public abstract void save(TAuthorityRbrelation transientInstance);
