+++ src/person/daizhongde/authority/hibernate/pojo/TAuthorityUser_NDid.java
@@ -16,7 +16,7 @@ public class TAuthorityUser_NDid implements java.io.Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 5157744434707155887L;
-	private Integer NUid;
+	private Long NUid;
 	private TAuthorityInst TAuthorityInst;
 //	private Integer NIid;
 	private Integer NDid;
@@ -26,7 +26,7 @@ public class TAuthorityUser_NDid implements java.io.Serializable {
 	private String CUsex;
 	
 	private String EmployeeNumber;
-	private Integer SupervisorId;
+	private Long SupervisorId;
 	
 	private String CUphone;
 	private String CUemail;
@@ -60,7 +60,7 @@ public class TAuthorityUser_NDid implements java.io.Serializable {
 	}
 
 	/** minimal constructor */
-	public TAuthorityUser_NDid(Integer NUid, TAuthorityInst TAuthorityInst,
+	public TAuthorityUser_NDid(Long NUid, TAuthorityInst TAuthorityInst,
 			String CUlogname, String CUname, String CUpassword, String CUsex,
 			String EmployeeNumber,
 			Date CUctime) {
@@ -75,9 +75,9 @@ public class TAuthorityUser_NDid implements java.io.Serializable {
 	}
 
 	/** full constructor */
-	public TAuthorityUser_NDid(Integer NUid, TAuthorityInst TAuthorityInst,
+	public TAuthorityUser_NDid(Long NUid, TAuthorityInst TAuthorityInst,
 			String CUlogname, String CUname, String CUpassword, String CUsex,
-			String EmployeeNumber,Integer SupervisorId,
+			String EmployeeNumber,Long SupervisorId,
 			String CUphone, String CUemail, String CUqq, String CUaddr,String CUnote,
 			String CUtel, String CUfax, Integer NDid, Date CUctime,
 			String CUcip, String CUcreator, Date CUmtime, String CUmip,
@@ -117,11 +117,11 @@ public class TAuthorityUser_NDid implements java.io.Serializable {
 
 	// Property accessors
 
-	public Integer getNUid() {
+	public Long getNUid() {
 		return this.NUid;
 	}
 
-	public void setNUid(Integer NUid) {
+	public void setNUid(Long NUid) {
 		this.NUid = NUid;
 	}
 
@@ -173,11 +173,11 @@ public class TAuthorityUser_NDid implements java.io.Serializable {
 		EmployeeNumber = employeeNumber;
 	}
 
-	public Integer getSupervisorId() {
+	public Long getSupervisorId() {
 		return SupervisorId;
 	}
 
-	public void setSupervisorId(Integer supervisorId) {
+	public void setSupervisorId(Long supervisorId) {
 		SupervisorId = supervisorId;
 	}
 	public String getCUphone() {
